+++ roman_discovery/matchers.py
@@ -43,6 +43,27 @@ class MatchByType:
 
 
 @dataclass
+class MatchBySubclass:
+    """Object matcher that select classes that are subclasses of a given type.
+
+    Constructor accepts a type or a tuple of types. E.g., the following instance will
+    find all Django models in a model.
+
+        from django.db import models
+        matcher = MatchBySubclass(models.Model)
+    """
+
+    object_type: Type
+
+    def __call__(self, obj: Any):
+        return (
+            inspect.isclass(obj)
+            and issubclass(obj, self.object_type)
+            and obj != self.object_type
+        )
+
+
+@dataclass
 class MatchByAttribute:
     """Object matcher that selects having an attribute with the given name.