+++ tests/test_matchers.py
@@ -1,6 +1,10 @@
 import pytest
 
-from roman_discovery.matchers import MatchByCallableAttribute, MatchByPattern
+from roman_discovery.matchers import (
+    MatchByCallableAttribute,
+    MatchByPattern,
+    MatchBySubclass,
+)
 
 
 @pytest.mark.parametrize(
@@ -30,4 +34,25 @@ def test_match_by_callable_attribute_should_return_relevant_objects(
     attribute, obj, result
 ):
     matcher = MatchByCallableAttribute(attribute)
+    assert matcher(obj) == result
+
+
+class StringSubclass(str):
+    pass
+
+
+@pytest.mark.parametrize(
+    "obj, result",
+    [
+        # instances don't match
+        (1, False),
+        ("foo", False),
+        # the class itself doesn't match
+        (str, False),
+        # subclasses match
+        (StringSubclass, True),
+    ],
+)
+def test_match_by_subclass_should_return_relevant_objects(obj, result):
+    matcher = MatchBySubclass(str)
     assert matcher(obj) == result